# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PokéVault API — private Pokémon card collection and resale tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sealed-products:
    get:
      tags: [SealedProducts]
      operationId: listSealedProducts
      summary: List all sealed products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: productType
          in: query
          required: false
          schema:
            type: string
        - name: set
          in: query
          required: false
          schema:
            type: string
        - name: condition
          in: query
          required: false
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, set, createdAt, quantity, productType]
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of sealed products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SealedProduct'
    post:
      tags: [SealedProducts]
      operationId: createSealedProduct
      summary: Create a sealed product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SealedProductInput'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealedProduct'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /sealed-products/{id}:
    get:
      tags: [SealedProducts]
      operationId: getSealedProduct
      summary: Get a single sealed product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealedProductDetail'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [SealedProducts]
      operationId: updateSealedProduct
      summary: Update a sealed product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SealedProductUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealedProduct'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [SealedProducts]
      operationId: deleteSealedProduct
      summary: Delete a sealed product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # Sealed Product Images
...
  /pickups:
    get:
      tags: [Pickups]
      operationId: listPickups
      summary: List all local pickups
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of pickups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pickup'
    post:
      tags: [Pickups]
      operationId: createPickup
      summary: Create a local pickup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PickupInput'
      responses:
        '201':
          description: Pickup created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pickup'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /pickups/{id}:
    get:
      tags: [Pickups]
      operationId: getPickup
      summary: Get a single pickup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pickup'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Pickups]
      operationId: updatePickup
      summary: Update a pickup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PickupUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pickup'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Pickups]
      operationId: deletePickup
      summary: Delete a pickup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # Export
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth
...
    ItemImage:
      type: object
      required: [id, objectPath, label, sortOrder, createdAt]
      properties:
        id:
          type: integer
        objectPath:
          type: string
        label:
          type: string
          description: "front | back | damage | receipt | sealed | other"
        caption:
          type: [string, 'null']
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    SealedProduct:
      type: object
      required: [id, name, productType, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        productType:
          type: string
          enum: [ETB, booster_box, booster_pack, tin, binder, case, other]
        set:
          type: [string, 'null']
        quantity:
          type: integer
        condition:
          type: [string, 'null']
        purchasePrice:
          type: [number, 'null']
        marketValue:
          type: [number, 'null']
        askingPrice:
          type: [number, 'null']
        soldPrice:
          type: [number, 'null']
        soldDate:
          type: [string, 'null']
          format: date
        status:
          type: string
          enum: [collection, for_sale, sold, wishlist, trade_binder, grading_pile]
        purchaseDate:
          type: [string, 'null']
          format: date
        purchaseSource:
          type: [string, 'null']
        tradeNotes:
          type: [string, 'null']
        location:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        coverImagePath:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, 'null']
          format: date-time
    SealedProductDetail:
      allOf:
        - $ref: '#/components/schemas/SealedProduct'
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: '#/components/schemas/ItemImage'
    SealedProductInput:
      type: object
      required: [name, productType]
      properties:
        name:
          type: string
          minLength: 1
        productType:
          type: string
          enum: [ETB, booster_box, booster_pack, tin, binder, case, other]
        set:
          type: string
        quantity:
          type: integer
          minimum: 1
        condition:
          type: string
        purchasePrice:
          type: number
          minimum: 0
        marketValue:
          type: number
          minimum: 0
        askingPrice:
          type: number
          minimum: 0
        status:
          type: string
          enum: [collection, for_sale, sold, wishlist, trade_binder, grading_pile]
        location:
          type: string
        notes:
          type: string
        purchaseDate:
          type: string
          format: date
        purchaseSource:
          type: string
        tradeNotes:
          type: string
    SealedProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        productType:
          type: string
          enum: [ETB, booster_box, booster_pack, tin, binder, case, other]
        set:
          type: string
        quantity:
          type: integer
          minimum: 1
        condition:
          type: string
        purchasePrice:
          type: number
          minimum: 0
        marketValue:
          type: number
          minimum: 0
        askingPrice:
          type: number
          minimum: 0
        soldPrice:
          type: number
          minimum: 0
        soldDate:
          type: string
          format: date
        status:
          type: string
          enum: [collection, for_sale, sold, wishlist, trade_binder, grading_pile]
        location:
          type: string
        notes:
          type: string
        purchaseDate:
          type: string
          format: date
        purchaseSource:
          type: string
        tradeNotes:
          type: string
        coverImagePath:
          type: string

    # Dashboard
...
    Pickup:
      type: object
      required: [id, buyerName, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        buyerName:
          type: string
        customerId:
          type: [integer, 'null']
        location:
          type: [string, 'null']
        meetingDatetime:
          type: [string, 'null']
          format: date-time
        itemDescription:
          type: [string, 'null']
        amountDue:
          type: [number, 'null']
        amountPaid:
          type: [number, 'null']
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PickupInput:
      type: object
      required: [buyerName]
      properties:
        buyerName:
          type: string
          minLength: 1
        customerId:
          type: integer
        location:
          type: string
        meetingDatetime:
          type: string
          format: date-time
        itemDescription:
          type: string
        amountDue:
          type: number
          minimum: 0
        amountPaid:
          type: number
          minimum: 0
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
          default: scheduled
        notes:
          type: string
    PickupUpdate:
      type: object
      properties:
        buyerName:
          type: string
          minLength: 1
        customerId:
          type: integer
        location:
          type: string
        meetingDatetime:
          type: string
          format: date-time
        itemDescription:
          type: string
        amountDue:
          type: number
          minimum: 0
        amountPaid:
          type: number
          minimum: 0
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: string
...